rename: useSignUp -> useTrySignUp #45
